Move all files in dedicate to the root folder
