--- tools/hermit_proxy/src/dedicate/mod.rs
@@ -1,4 +1,0 @@
-pub mod qemu;
-pub mod multi;
-pub mod proto;
-pub mod socket;
